Trying to completely remove the BlazorWasmDebuggingExtension and getting URL information from launchSettings.json
