--- src/Razor/src/Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension/.npmrc
@@ -1,2 +1,0 @@
-﻿# Auto generated file from Gardener Plugin CentralFeedServiceAdoptionPlugin
-registry=https://pkgs.dev.azure.com/dnceng/public/_packaging/dotnet-public-npm/npm/registry/
