--- src/Razor/src/Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension/.yarnrc
@@ -1,1 +1,0 @@
-﻿--ignore-engines true
