--- src/Razor/src/Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension/README.md
@@ -1,7 +1,0 @@
-﻿# Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension
-
-This is a companion extension to the [C# extension](https://marketplace.visualstudio.com/items?itemName=ms-dotnettools.csharp) that brings some improvements to the experience for debugging Blazor WebAssembly in VS Code. This extension should **not** be installed by itself and will automatically install the C# extension.
-
-For more information on Blazor, visit the [Blazor homepage](https://blazor.net).
-
-For more information on debugging Blazor WebAssembly applications, vist our official documentation page on [debugging Blazor WASM](https://docs.microsoft.com/aspnet/core/blazor/debug).
