--- src/Razor/src/Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension/package.json
@@ -1,67 +1,0 @@
-{
-  "name": "blazorwasm-companion",
-  "displayName": "Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension",
-  "description": "A companion extension for debugging Blazor WebAssembly applications in VS Code. Must be installed alongside the C# extension.",
-  "version": "1.1.4",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dotnet/razor.git"
-  },
-  "license": "MIT",
-  "engines": {
-    "vscode": "^1.69.0"
-  },
-  "icon": "images/blazorIcon.png",
-  "dotnetRuntimeVersion": "6.0",
-  "debugProxyVersion": "6.0.1",
-  "categories": [
-    "Other"
-  ],
-  "extensionKind": [
-    "ui"
-  ],
-  "publisher": "ms-dotnettools",
-  "activationEvents": [
-    "onCommand:blazorwasm-companion.launchDebugProxy"
-  ],
-  "main": "./dist/extension.js",
-  "extensionDependencies": [
-    "ms-dotnettools.vscode-dotnet-runtime",
-    "ms-dotnettools.csharp"
-  ],
-  "capabilities": {
-    "untrustedWorkspaces": {
-      "supported": true
-    }
-  },
-  "contributes": {},
-  "scripts": {
-    "vscode:prepublish": "yarn run build",
-    "clean": "rimraf dist",
-    "build": "yarn run clean && yarn run lint && yarn run compile",
-    "compile": "tsc -p ./",
-    "watch": "tsc -watch -p ./",
-    "pretest": "yarn run compile && yarn run lint",
-    "lint": "tslint --project ./"
-  },
-  "devDependencies": {
-    "@types/glob": "^7.1.3",
-    "@types/mocha": "^8.0.4",
-    "@types/node": "^14.18.12",
-    "@types/vscode": "^1.69.0",
-    "@typescript-eslint/eslint-plugin": "^4.9.0",
-    "@typescript-eslint/parser": "^4.9.0",
-    "extract-zip": "^2.0.1",
-    "glob": "^7.1.6",
-    "minimatch": "3.0.5",
-    "mocha": "^8.1.3",
-    "node-fetch": "^2.6.7",
-    "rimraf": "2.6.3",
-    "tslint": "^5.11.0",
-    "typescript": "^4.5.4",
-    "vscode-test": "^1.4.1"
-  },
-  "dependencies": {
-    "nanoid": "^4.0.0"
-  }
-}
