--- src/Razor/src/Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension/tsconfig.json
@@ -1,9 +1,0 @@
-{
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-	  "outDir": "dist"
-	},
-	"include": [
-	  "./src/**/*"
-	]
-}
