--- src/Razor/src/Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension/tslint.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../tslint.json"
-}
