+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode.Extension/Microsoft.AspNetCore.Razor.VSCode.Extension.npmproj
@@ -14,12 +14,6 @@
       SkipGetTargetFrameworkProperties="true"
       UndefineProperties="TargetFramework"
       Private="false" />
-    <ProjectReference
-      Include="..\Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension\Microsoft.AspNetCore.Razor.VSCode.BlazorWasmDebuggingExtension.npmproj"
-      ReferenceOutputAssemblies="false"
-      SkipGetTargetFrameworkProperties="true"
-      UndefineProperties="TargetFramework"
-      Private="false" />
   </ItemGroup>
 
   <ItemGroup>
